invoice-ingest/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.27.2
supabase==2.7.4
google-cloud-storage==2.17.0
